Move TaskScheduler and TaskSchedulerService to org.example.core.service.task.scheduler pachage
